// src/app/generated/models/review-case-detail.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Advisory } from '../models/advisory';
import { AnalysisDetail } from '../models/analysis-detail';
import { CaseDetail } from '../models/case-detail';
import { Remedy } from '../models/remedy';
import { ReviewTask } from '../models/review-task';
export interface ReviewCaseDetail {
  analysis: AnalysisDetail;
  case: CaseDetail;
  farmerName?: string;
  priorAdvisory?: Advisory;
  suggestedDiseaseId?: string | null;
  suggestedRemedies: Array<Remedy>;
  task: ReviewTask;
}